Moved 1.13 only MaterialHook Methods to MaterialHelper (for whatever reason github doesn't like me reverting the last commit - so this is a followup commit)
