--- src/me/mrCookieSlime/CSCoreLibPlugin/compatibility/MaterialHook13.java
@@ -1,167 +1,0 @@
-package me.mrCookieSlime.CSCoreLibPlugin.compatibility;
-
-import org.bukkit.Material;
-
-/**
- */
-public class MaterialHook13 {
-
-    static Material getSaplingFromLog(Material log) {
-        switch (log) {
-            case OAK_LOG:
-                return Material.OAK_SAPLING;
-            case SPRUCE_LOG:
-                return Material.SPRUCE_SAPLING;
-            case BIRCH_LOG:
-                return Material.BIRCH_SAPLING;
-            case JUNGLE_LOG:
-                return Material.JUNGLE_SAPLING;
-            case ACACIA_LOG:
-                return Material.ACACIA_SAPLING;
-            case DARK_OAK_LOG:
-                return Material.DARK_OAK_SAPLING;
-        }
-        return Material.OAK_SAPLING;
-    }
-
-
-    static Material getWoodFromLog(Material log) {
-        switch (log) {
-            case OAK_LOG:
-                return Material.OAK_WOOD;
-            case SPRUCE_LOG:
-                return Material.SPRUCE_WOOD;
-            case BIRCH_LOG:
-                return Material.BIRCH_WOOD;
-            case JUNGLE_LOG:
-                return Material.JUNGLE_WOOD;
-            case ACACIA_LOG:
-                return Material.ACACIA_WOOD;
-            case DARK_OAK_LOG:
-                return Material.DARK_OAK_WOOD;
-        }
-        return Material.OAK_WOOD;
-    }
-
-    static boolean isLog(Material log) {
-        switch (log) {
-            case OAK_LOG:
-            case SPRUCE_LOG:
-            case BIRCH_LOG:
-            case JUNGLE_LOG:
-            case ACACIA_LOG:
-            case DARK_OAK_LOG:
-                return true;
-        }
-        return false;
-    }
-
-
-    static Material[] initWoolColours() {
-        return new Material[]{
-                Material.WHITE_WOOL,
-                Material.ORANGE_WOOL,
-                Material.MAGENTA_WOOL,
-                Material.LIGHT_BLUE_WOOL,
-                Material.YELLOW_WOOL,
-                Material.LIME_WOOL,
-                Material.PINK_WOOL,
-                Material.GRAY_WOOL,
-                Material.LIGHT_GRAY_WOOL,
-                Material.CYAN_WOOL,
-                Material.PURPLE_WOOL,
-                Material.BLUE_WOOL,
-                Material.BROWN_WOOL,
-                Material.GREEN_WOOL,
-                Material.RED_WOOL,
-                Material.BLACK_WOOL
-        };
-    }
-
-
-    static Material[] initStainedGlassColours() {
-        return new Material[]{
-                Material.WHITE_STAINED_GLASS,
-                Material.ORANGE_STAINED_GLASS,
-                Material.MAGENTA_STAINED_GLASS,
-                Material.LIGHT_BLUE_STAINED_GLASS,
-                Material.YELLOW_STAINED_GLASS,
-                Material.LIME_STAINED_GLASS,
-                Material.PINK_STAINED_GLASS,
-                Material.GRAY_STAINED_GLASS,
-                Material.LIGHT_GRAY_STAINED_GLASS,
-                Material.CYAN_STAINED_GLASS,
-                Material.PURPLE_STAINED_GLASS,
-                Material.BLUE_STAINED_GLASS,
-                Material.BROWN_STAINED_GLASS,
-                Material.GREEN_STAINED_GLASS,
-                Material.RED_STAINED_GLASS,
-                Material.BLACK_STAINED_GLASS
-        };
-    }
-
-
-    static Material[] initStainedGlassPaneColours() {
-        return new Material[]{
-                Material.WHITE_STAINED_GLASS_PANE,
-                Material.ORANGE_STAINED_GLASS_PANE,
-                Material.MAGENTA_STAINED_GLASS_PANE,
-                Material.LIGHT_BLUE_STAINED_GLASS_PANE,
-                Material.YELLOW_STAINED_GLASS_PANE,
-                Material.LIME_STAINED_GLASS_PANE,
-                Material.PINK_STAINED_GLASS_PANE,
-                Material.GRAY_STAINED_GLASS_PANE,
-                Material.LIGHT_GRAY_STAINED_GLASS_PANE,
-                Material.CYAN_STAINED_GLASS_PANE,
-                Material.PURPLE_STAINED_GLASS_PANE,
-                Material.BLUE_STAINED_GLASS_PANE,
-                Material.BROWN_STAINED_GLASS_PANE,
-                Material.GREEN_STAINED_GLASS_PANE,
-                Material.RED_STAINED_GLASS_PANE,
-                Material.BLACK_STAINED_GLASS_PANE
-        };
-    }
-
-
-    static Material[] initTerracottaColours() {
-        return new Material[]{
-                Material.WHITE_TERRACOTTA,
-                Material.ORANGE_TERRACOTTA,
-                Material.MAGENTA_TERRACOTTA,
-                Material.LIGHT_BLUE_TERRACOTTA,
-                Material.YELLOW_TERRACOTTA,
-                Material.LIME_TERRACOTTA,
-                Material.PINK_TERRACOTTA,
-                Material.GRAY_TERRACOTTA,
-                Material.LIGHT_GRAY_TERRACOTTA,
-                Material.CYAN_TERRACOTTA,
-                Material.PURPLE_TERRACOTTA,
-                Material.BLUE_TERRACOTTA,
-                Material.BROWN_TERRACOTTA,
-                Material.GREEN_TERRACOTTA,
-                Material.RED_TERRACOTTA,
-                Material.BLACK_TERRACOTTA
-        };
-    }
-
-    static Material[] initGlazesTerracottaColours() {
-        return new Material[]{
-                Material.WHITE_GLAZED_TERRACOTTA,
-                Material.ORANGE_GLAZED_TERRACOTTA,
-                Material.MAGENTA_GLAZED_TERRACOTTA,
-                Material.LIGHT_BLUE_GLAZED_TERRACOTTA,
-                Material.YELLOW_GLAZED_TERRACOTTA,
-                Material.LIME_GLAZED_TERRACOTTA,
-                Material.PINK_GLAZED_TERRACOTTA,
-                Material.GRAY_GLAZED_TERRACOTTA,
-                Material.LIGHT_GRAY_GLAZED_TERRACOTTA,
-                Material.CYAN_GLAZED_TERRACOTTA,
-                Material.PURPLE_GLAZED_TERRACOTTA,
-                Material.BLUE_GLAZED_TERRACOTTA,
-                Material.BROWN_GLAZED_TERRACOTTA,
-                Material.GREEN_GLAZED_TERRACOTTA,
-                Material.RED_GLAZED_TERRACOTTA,
-                Material.BLACK_GLAZED_TERRACOTTA
-        };
-    }
-}
